fix: make image upload predictable
Initially, the image upload function would get called immediately even
when the button was not clicked. Here was why:

So the issue arises from the way the `uploadImage` function is passed in the
`onClick` prop of the upload button. How onClick basically works is that
it accepts a function definition and then carries out that function definition when the
the click event occurs. What we were passing initially was actually the
returned value from calling `uploadImage` and not it's function definition.
So when the button gets clicked, `onClick` would try to call that returned
value but since `uploadImage` does not return anything, nothing happens
and we just get a really buggy behaviour.