feat: improve drag and drop behaviour
Previously, children nodes that can be nested within other similar
parent nodes cannot be dropped on the parent node itself, but must be
arranged among the other children nodes of that parent node. This commit
can allow the children node to be dropped in the parent node itself, and
the children node will be added in automatically.