Show tree structure of the navigation menu (#52)
* Add atlas kit tree package

* Add front-end endpoint for tree test

* Set up sample code for tree

* [WIP] feat: include drag & drop functionality

You can now drag and drop item across one another, however there are no
rules placed to restrict this behaviour, which is required
for our CMS. This will be addressed in a later commit.

* Rename tree component, add utils to flatten tree

* [WIP]feat: impose ordering rules

Currently there are no restrictions imposed on how the tree items can be
reordered.

* feat: link tree building with backend

This commit removes the need for a static data to be in place for the
tree to be constructed. It calls the backend for the hierarchy
structure.

* Convert typescript utils into javascript

* Add navigation and unlinked pages as top-level ancestors to the drag and drop tree

* Reorganized utils files to have a general utils folder for all components

* Reorganize code into a separate general utils folder, create a higher order component to wrap tree elements in draggable props

* Modify tree to accept unlinked collections and resources

* fix: rename nav & unlinked pages node type

This commit fixes the broken drag & drop rules that were initially in
place for collection dragging behaviour

* feat: disable drag & drop for top level nodes

This commit disables drag & drop for the top level nodes: `navigation` and `unlinked pages`.

* feat: allow expanded nodes to be dragged

This commit fixes the height of the entire tree view so that expanded
nodes can be dragged. The reason why it wasn't working previously was
because the height of the tree view was not fixed.

* feat: improve drag and drop behaviour

Previously, children nodes that can be nested within other similar
parent nodes cannot be dropped on the parent node itself, but must be
arranged among the other children nodes of that parent node. This commit
can allow the children node to be dropped in the parent node itself, and
the children node will be added in automatically.