fix: retrieve leftNavPages in EditCollectionPage
For context, when the user clicks on the collection page under the
`collections/:collectionName` endpoint, its sibling collection pages
gets passed into `location` of the router. The `EditCollectionPage` then
retrieves those pages and renders them  in the leftnav of the preview.

Since we are flattening the collections tab into the pages tab, we
currently do not keep track of the sibling collection pages of any
collection page. Hence this commit moves the logic to find the sibling
collection pages into `EditCollectionPage`.