BREAKING CHANGE: Include collection pages in the `Pages` tab (#64)
* feat!: show collection pages in the pages tab

THIS WILL ONLY WORK WITH THE BACKEND BRANCH
`feat/flatten-collection-into-pages`.

This commit now introduces a new settings modal to handle CRUD
operations for collection pages and displays collection pages in the
`Pages` tab.

* feat: remove collections from sidebar

Since collection pages are shown on the `pages` tab, there is no need
for a separate `Collections` tab to show the various collection pages.

* fix: retrieve leftNavPages in EditCollectionPage

For context, when the user clicks on the collection page under the
`collections/:collectionName` endpoint, its sibling collection pages
gets passed into `location` of the router. The `EditCollectionPage` then
retrieves those pages and renders them  in the leftnav of the preview.

Since we are flattening the collections tab into the pages tab, we
currently do not keep track of the sibling collection pages of any
collection page. Hence this commit moves the logic to find the sibling
collection pages into `EditCollectionPage`.