fix: prevent frontend from crashing during hasErrors check
hasErrors calls hasSectionErrors, which iterates across all form fields to check for errors. When a user toggles between Hero Dropdown and Button, the hidden form fields' error objects are set to undefined. However, this crashes when hasSectionErrors is computed. The fix is to set the hidden form fields' error objects to an empty str