fix: render list items based on the node's type
The rendering of the item should depend on the type of the node rather than
checking if the node has children nodes. That way, if a `collection` type
node does not have any children, it will still render with a folder icon
and not a page icon.

This commit makes sure that the list item is rendered based on its type.