feat: include route to edit main menu (#74)
* feat: include EditNav route

The `EditTree` component has been moved and is now called the `EditNav`
component.

* fix: correct link to editnav route

Link in the Menu Component to the `EditNav` route was missing a few
characters. This commit fixed that

* fix: render list items based on the node's type

The rendering of the item should depend on the type of the node rather than
checking if the node has children nodes. That way, if a `collection` type
node does not have any children, it will still render with a folder icon
and not a page icon.

This commit makes sure that the list item is rendered based on its type.

* feat: improve drag and drop behaviour rules

This commit makes the code slightly more readable. Hopefully.

* style: clean up code in EditNav