fix: add collapsed child nodes in flattened tree
In the tree flattening function (that we ripped out of
`@atlaskit/tree`), it's default behaviour would remove children nodes
under a parent node that's been collapsed. The reason behind why it did
that was because that function was used to render the tree, hence it did
not need the child nodes under a parent that has been collapsed.

This was causing crashing issues when we needed to render the navigation
preview. The fix was to include child nodes that were under collapsed
parent nodes.