feat: add navigation preview (#86)
* Feat/read tree (#82)

* feat: create function that reads tree to create nav bar

* refactor: move tree util functions to utils folder

* [WIP]feat: include navigation preview in EditNav

This commit is not completed as the nav preview does not update after
drag and drop changes are made.

* fix: add collapsed child nodes in flattened tree

In the tree flattening function (that we ripped out of
`@atlaskit/tree`), it's default behaviour would remove children nodes
under a parent node that's been collapsed. The reason behind why it did
that was because that function was used to render the tree, hence it did
not need the child nodes under a parent that has been collapsed.

This was causing crashing issues when we needed to render the navigation
preview. The fix was to include child nodes that were under collapsed
parent nodes.

* style: remove hamburger and menu item selector

These two items are buggy right now, hence they are commented out. The
item selector line is not aligned properly either when navigation item
is hovered upon.

* feat: prevent nav item link redirection on click

This commit prevents the anchor tags from redirecting the user after
they click on it.

* feat: add support for false collections