fix: remove media color title from consideration in retrieving errors
We are currently using the title for media colors as its key when
rendering the media color fields. This means that the media color field
titles cannot be empty even at the first render.

However, this disables the save button since our retrieve error
function checks for whether any of the fields in the errors object
is truthy, and the media color fields title is not empty (and thus
truthy).

By deleting the media color fields' titles, we get the correct
behavior on the save button.