[Feat] Create settings page (#99)
* feat: Add settings route to the app

* feat: Add a barebones settings page

* feat: add new form field components

* feat: store on state and render all necessary state values

* feat: add white background to settings container

* feat: add a heading line for social media fields

* feat: add two new form field types for settings page

This commit introduces two new form fields, FormFieldHorizontal and
FormFieldImage, that allow the title of the field to be next to the
text input.

FormFieldImage will be fleshed out in future commits to include a
color picker.

* feat: add color boxes next to text input

* style: add styles to form fields

This commit adds a color box to the FormFieldColor component and
adjusts the spacing between the titles and text inputs

* feat: add save functionality to the settings page

* style: add margin to FormFieldHorizontal

* fix: set state properly and save settings properly

This commit accomplishes the following:
- decode base64 content received from backend
- repair set state functionality in this.changeHandler
- general cleanup

* feat: switch horizontal form fields to flexbox

* style: align horizontal form fields

* refactor: introduce ColorPicker component and toggle

This commit isolates the ColorPicker code chunk as a component and
introduces a toggle in Settings that will toggle the ColorPicker
when it is converted to a modal.

* feat: flesh out colorPickerToggle

* feat: add color picking package

* style: halven color field width

* feat: complete ColorPicker functionality

This commit achieves the following:
- complete functionality of ColorPicker modal
- refactor ColorPicker toggling into activation and disabling
handlers
- introduce event listeners so that ColorPicker modal can be
disabled by an escape key or a mouseclick outside of the modal

* feat: reflect color changes in real time

* style: reposition save button to bottom-right

* feat: add validation for social media fields

* fix: delete errors from config settings

* fix: social media validation

previously, if a social media field is made empty (after deleting
some characters), an error would show up. this commit ensures that
errors do not show up on empty fields.

* feat: prevent saving if errors exist

this commit introduces a `retrieveErrors` function which reduces
the `errors` object into an array of all error values, and checks
if there are any errors.

with this function, we can disable the 'Save' button if errors
exist

* fix: remove validation for color fields

* fix: add keys to React lists

* fix: remove media color title from consideration in retrieving errors

We are currently using the title for media colors as its key when
rendering the media color fields. This means that the media color field
titles cannot be empty even at the first render.

However, this disables the save button since our retrieve error
function checks for whether any of the fields in the errors object
is truthy, and the media color fields title is not empty (and thus
truthy).

By deleting the media color fields' titles, we get the correct
behavior on the save button.

* style: style select button on color picker

* style: rearrange fields and style section header text

* refactor: accomodate changes in the backend

This commit modifies the frontend to accomodate changes to the
response received in the backend. Now, we only receive the relevant
fields from _config.yml as opposed to all fields. We also no longer
receive the _config.yml file's sha

* feat: allow social media links to end with /

* refactor: rename variables

* fix: simplify error detection