feat: upload image after naming
Previously, the image file is uploaded upon clicking the 'Upload image'
button and the preview image in the ImageSettingsModal is grabbed from
the already uploaded image. However, this is inefficient because when
a user wants to rename a newly uploaded image, two API calls are required.

This commit reorders the upload operation so that the image file is only
uploaded AFTER the user chooses the file name. This reduces the number of
API calls being made to 1.