Feat/upload image rename modal (#107)
* feat: allow for 4-letter image extensions in validation

* feat: add modal to allow renaming of image after upload

* fix: remove unused Media tab

* fix: remove mention of Media tab from router

* fix: remove mention of Media tab from router

* fix: remove unnecessary image rename modal

the ImageUploadModal was essentially the same as the ImageSettingsModal
so we are reusing ImageSettingsModal instead

* feat: upload image after naming

Previously, the image file is uploaded upon clicking the 'Upload image'
button and the preview image in the ImageSettingsModal is grabbed from
the already uploaded image. However, this is inefficient because when
a user wants to rename a newly uploaded image, two API calls are required.

This commit reorders the upload operation so that the image file is only
uploaded AFTER the user chooses the file name. This reduces the number of
API calls being made to 1.

* fix: requested changes from review

* fix: remove unnecessary code