feat: update ImageSettingsModal
This commit achieves two things:

1. It allows ImageSettingsModal to receive a `toReload` prop since
we do not always want the page to reload. In the case of EditPage
with the ImageModal, we want users to be able to upload an image
and then select it right away. If we do not have a `toReload` prop,
the window will reload and the ImageModal will be reset.

2. It adds custom logic for `isPendingUpload`. In cases where the
image/file is pending upload, there should not be a delete button
in the ImageSettingsModal, and the Save button would not be subject
to the usual constraints (requiring a `sha` value before the button
becomes active since the create operation does not require a `sha`)