feat: refactor ImagesModal to be a PureComponent
This commit accomplishes a couple of things:

1. It introduces the functionality to rename image before uploading for the
ImageModal (this functionality was firrst introduced for the Images tab in
PR #107).

2. It refactors the ImagesModal to be a PureComponent. Previously, the ImageModal
retrieved images and saved this images in state. This functionality can be moved
to its parent component, EditPage, so that only an `images` prop is passed to the
ImageModal.

This is beneficial because the ImageModal logic is simple and its output really
depends on the images that are passed as props to it. Since the key for the images
are the filenames, this means that when we upload a new image, React will perform
a shallow compare based on the filenames and only render the new uploaded image
instead of re-rendering all the images.