[feat] Allow rename before image upload in EditPage, refactor ImageModal (#117)
* feat: update ImageSettingsModal

This commit achieves two things:

1. It allows ImageSettingsModal to receive a `toReload` prop since
we do not always want the page to reload. In the case of EditPage
with the ImageModal, we want users to be able to upload an image
and then select it right away. If we do not have a `toReload` prop,
the window will reload and the ImageModal will be reset.

2. It adds custom logic for `isPendingUpload`. In cases where the
image/file is pending upload, there should not be a delete button
in the ImageSettingsModal, and the Save button would not be subject
to the usual constraints (requiring a `sha` value before the button
becomes active since the create operation does not require a `sha`)

* feat: refactor ImagesModal to be a PureComponent

This commit accomplishes a couple of things:

1. It introduces the functionality to rename image before uploading for the
ImageModal (this functionality was firrst introduced for the Images tab in
PR #107).

2. It refactors the ImagesModal to be a PureComponent. Previously, the ImageModal
retrieved images and saved this images in state. This functionality can be moved
to its parent component, EditPage, so that only an `images` prop is passed to the
ImageModal.

This is beneficial because the ImageModal logic is simple and its output really
depends on the images that are passed as props to it. Since the key for the images
are the filenames, this means that when we upload a new image, React will perform
a shallow compare based on the filenames and only render the new uploaded image
instead of re-rendering all the images.