[Feat - EditPage, ImagesModal] Highlight selected image (#124)
* style: linting and props validation

* fix: allow same file to be uploaded consecutively

Currently in the ImagesModal or Images tab, if you click on Add New Image and select a file, then
click exit on the ImageSettingsModal, you will not be able to select that image again - the file
reader simply doesn't pop up. By clearing the target value on click, the onChange event will be
triggered, whereas before, it does not get triggered since there is no change in target value

* fix: remove edit icon from media card

* style: lint EditPAge

* style: lint

* feat: introduce image highlight and selection functionality

This commit changes the flow for image selection in EditPage. Instead of selecting the image on
click, when a user clicks on an image in the ImagesModal, the image is first highlighted and the
image must be selected using a Select button on the top right side

* style: position select button

* fix: apply fix to #125 for Images component