refactor: replace type-specified media components
We were declaring components (`ImageCard`/`FileCard`/`ImageUploadCard`, etc.) that were constructed from the corresponding base media component, but with the type specified.

This proved to be a problem as when we include a new prop for the base
media component, we would then have to include that prop for the
type-specified component and make sure that new prop is passed into the
base media component it was created from.

It can get pretty messy as there are now multiple points to update the
code, hence it'd be better to just use the base media components when
rendering and specify the type there instead of creating a component
that builds on top of that base, but with the type specified.