fix(resources): re-enable creation of new resources
This fixes a bug introduced in commit e593255 (a commit to show the resource settings modal only
after the data is fetched for the selected resource). The condition to show the modal was only after the `sha`
of the resource was fetched. I forgot to account for the fact that the resource settings modal is also
used when creating a new resource. Since new resources have no `sha`, the modal doesn't show. To fix
that, the condition to show the modal checks for either the `sha` of the fetched resource or checks if the `isNewPost` prop is true.