improvement: resources file url (#134)
* feat(resource-settings): introduce 'select file' button and disable url input

Users should not need to remember the permalinks of where their files are and input them. Hence this
branch will introduce commits for users to select their file from a modal and the permalink will be
added in for them.

* refactor(resource-settings): rename button click handler

* build: update packages

* feat(files-modal): implement files modal

* feat(files-modal): allow uploads through modal

* refactor((images)): switch to using base media components

Initially, media-related components were initialised into their type before they are used. (i.e
constructing `MediaCard` with type `image`, then assigning it to `ImageCard` and using `ImageCard`).
This was a problem as new props being passed through ImageCard in the `render` function had to be
updated in the `MediaCard` initialization call as well. Instead of updating two places, we should
just use the default `MediaCard` component

* fix(resource-settings): nest modal buttons into modalContent div to apply intended styling

The modal buttons div was not a child to the div with class `modalContent`. This caused the buttons
to not have the intended css styling. To fix it, it was made to be a child of the `modalContent`
div.

* feat(resource-settings): validate file selection

Validation is in place for file selection in the resource settings modal.

 The way validation is checked is slightly different for the FormFieldFile component. Usually, validation is done when the form field onChange handler is called. This is easily done since users can key into the input and invoke it.

Since the FormFieldFile input's value changes based on the file selected, the onChange handler isn't called at all. To mimic a similar onChange behaviour like other FormField components, `useEffect` is used to call `onFieldChange` whenever the input value changes.

* fix(resource-settings): erase file url validation error if user switches to permalink

* improvement: show modal only after data is fetched

Previously, the modal would load and the form fields would be empty.
Then the data is fetched and the fields are populated. This is slightly
trippy for users.

This commit now just loads the overlay first, and once the data is
fetchd to populate the modal, the modal then renders.

* fix: remove blue border around cards in images tab

Introduced a bug where a blue border appeared on all the image cards in the image tab.

This commit fixes it.

* refactor: replace type-specified media components

We were declaring components (`ImageCard`/`FileCard`/`ImageUploadCard`, etc.) that were constructed from the corresponding base media component, but with the type specified.

This proved to be a problem as when we include a new prop for the base
media component, we would then have to include that prop for the
type-specified component and make sure that new prop is passed into the
base media component it was created from.

It can get pretty messy as there are now multiple points to update the
code, hence it'd be better to just use the base media components when
rendering and specify the type there instead of creating a component
that builds on top of that base, but with the type specified.

* build: update package-lock.json

* fix(resources): re-enable creation of new resources

This fixes a bug introduced in commit e593255 (a commit to show the resource settings modal only
after the data is fetched for the selected resource). The condition to show the modal was only after the `sha`
of the resource was fetched. I forgot to account for the fact that the resource settings modal is also
used when creating a new resource. Since new resources have no `sha`, the modal doesn't show. To fix
that, the condition to show the modal checks for either the `sha` of the fetched resource or checks if the `isNewPost` prop is true.