fix: make entire PageCard component clickable
The Pages layout displays a list of PageCard components which act
as links towards individual EditPages for users to modify their
markdown content. However, currently, each PageCard component is
only clickable within a narrow band. When the cursor is on the edge
of the PageCard component, clicking on it does not bring you to the
EditPage to start editing.

The desired behavior is to ensure that the entire component is clickable.
To do this, we need to allow the Link component to make up the entire
<li> element.

This commit makes the necessary stylistic changes to allow the Link component
to take up the entire <li> component so the whole PageCard component is
clickable. Style changes were also made to the "Homepage" list element
to keep the Pages layout visually consistent.