fix: display sgds icons properly
This commit fixes issue #151. Previously, we had blindly copied the
sgds icons into our the isomer-template.scss file but we had not
checked to ensure that it was working. This installation of the SGDS
did not work, so the icons we tried to display were missing.

By following instructions from the sgds-govtech npm page, which can
be found here: https://www.npmjs.com/package/sgds-govtech, I was able
to successfully import the sgds icons and get them to display on the
homepage preview.

This commit also modifies the Homepage resources section button
alignment so that it's now centered.