Fix/misc ui bugs (#153)
* fix: display "Back to sites" button on Settings page

This commit makes the Settings page consistent with other pages by
enabling the "Back to sites" button for the Settings page

* fix: set default btn bg color to be white, center align btn icon

This commit accomplishes two things:
1. It center aligns button icons by removing the right padding and
sets "text-align: center" on the base button class

2. It sets the default button bg color to be white to blend in with
the CMS background.

* fix: make entire PageCard component clickable

The Pages layout displays a list of PageCard components which act
as links towards individual EditPages for users to modify their
markdown content. However, currently, each PageCard component is
only clickable within a narrow band. When the cursor is on the edge
of the PageCard component, clicking on it does not bring you to the
EditPage to start editing.

The desired behavior is to ensure that the entire component is clickable.
To do this, we need to allow the Link component to make up the entire
<li> element.

This commit makes the necessary stylistic changes to allow the Link component
to take up the entire <li> component so the whole PageCard component is
clickable. Style changes were also made to the "Homepage" list element
to keep the Pages layout visually consistent.

* fix: make all homepage preview elements non-clickable

This commit addresses issue #152 by removing the url values from all
homepage preview template components, and converting all <a href> tags
to <div> tags so that they're not actually clickable. For more details
on why we're doing this, refer to the description for issue #152.

* fix: display sgds icons properly

This commit fixes issue #151. Previously, we had blindly copied the
sgds icons into our the isomer-template.scss file but we had not
checked to ensure that it was working. This installation of the SGDS
did not work, so the icons we tried to display were missing.

By following instructions from the sgds-govtech npm page, which can
be found here: https://www.npmjs.com/package/sgds-govtech, I was able
to successfully import the sgds icons and get them to display on the
homepage preview.

This commit also modifies the Homepage resources section button
alignment so that it's now centered.