fix: replace native delete button with LoadingButton in image modal
This commit replaces the native html button element which we use for
deleting files/images with our LoadingButton component. The main difference
is that the button remains disabled until the file sha is retrieved - without
the SHA, the delete operation will fail (Github will respond with a 409)
and our modal will be stuck.