feat: combine files & images modal (#135)
* feat(edit-page): transfer media uploading responsibility to media modal

Previously, when a user invokes the image modal when editing a page and clicks on the `upload image
card`, the responsibility of uploading was shared between the `EditPage` component and the
`ImagesModal`. The `EditPage` component was responsible for invoking the `MediaSettingsModal` to
upload the image. This doesn't make sense as the responsibility of uploading should only be on the
modal that invoked it. The role of the media modal becomes ambiguous which is unnecessary. When
interacting with other components, all the media modal should do is provide the file/image to the
component that it is interacting with and not give it the responsibility of uploading any media
files. This commit makes it such that EditPage no longer has to take care of uploading media related
items and moves it into the media modal to do so. This makes the code cleaner.

* refactor(resource-settings): switch from files modal to media modal

* fix(media-setting): correct propType

* fix: replace native delete button with LoadingButton in image modal

This commit replaces the native html button element which we use for
deleting files/images with our LoadingButton component. The main difference
is that the button remains disabled until the file sha is retrieved - without
the SHA, the delete operation will fail (Github will respond with a 409)
and our modal will be stuck.

Co-authored-by: kwajiehao <jie_hao_kwa@alumni.brown.edu>