fix: allow settings button in PageCard to be clicked
Previously, the settings button in the PageCard was nested inside
the Link component's <a> tag, which is not allowed (interactive elements
are not allowed to be nested in each other, or the nested component
loses its interactivity) and resulted in the settings button being
unclickable.

This commit fixes that by moving the button element outside the Link
component.