fix: remove edit button from page settings modal
Previously, one needed to click on the "Edit" button before editing
the permalink of a given page. This is inconsistent with the experience
of editing the title on the page settings modal.

This commit edits the page settings modal so that the "Edit" button
is removed, and the permalink is edited in the same way as the title.

This commit also ensures that the "Save" and "Delete" buttons are
disabled until the file SHA is received, so that we don't attempt to
make API calls which fail (it is impossible to update a file or delete
a file without its current SHA)

This commit fixes issue #166.