fix: elaborate on filename validation
This commit breaks file name validation upon upload into two separate
steps:
1. a check for special characters
2. a check for proper file extensions

Previously, the two checks were made with the same regex expression,
which made it hard to provide a clear error message to the user. This
commit breaks them down so that users can have a clearer idea on how to
fix the error, with more elaborate error messages.