refactor: combine EditPage and EditResourcePage
Since most of the code between EditPage and EditResourcePage are duplicated,
we can add a few conditional statements to EditPage and delete EditResourcePage
altogether.

As part of this change, we need to pass an `isResourcePage` prop to EditPage so
that it knows whether to hit the pages endpoint, or the resources endpoint.