refactor: Header component to use state change for back navigation
Currently, the Header component contains an <a> tag with a link to
the previous page. This commit refactors the header component so that
instead of using an <a> tag to redirect the user, we redirect using
React router based on a state change. Upon clicking the "Back to pages"
or "Back to site" button, a toggler toggles the shouldPerformBackNav
state and the component redirects backwards.

This change is made in preparation for an "Are you sure you want to
navigate away?" modal when users are in the `EditPage` component and
try to navigate away when they have unsaved changes.