feat: add GenericWarningModal component
This commit adds the GenericWarningModal component, which performs in
a similar manner to the DeleteWarningModal, but allows the user to
specify the warning title and warning text.

This commit then modifies the Header component so that a user is
prompted to confirm that he/she wants to navigate away from the page
whenever they click the "Back to Pages" button, if the Header is part
of an EditPage.