feat: trigger warning modal only if page content has changed
This commit introduces a shouldAllowEditPageBackNav prop to the Header component,
to be used in conjunction with the isEditPage prop. The
shouldAllowEditPageBackNav prop determines whether the warning modal
should be triggered for an EditPage layout.

In this case, the original markdown content is stored upon download
and used to compare with the updated markdown editor value. The value
of shouldAllowEditPageBackNav is thus true if the original content matches
the editor content, and false otherwise. This allows us to achieve the behavior
of preventing user from navigating away from the page if they have unsaved changes.
This works because everytime the user saves, the page is reloaded so the new content
is retrieved.