Fix: fixed bug where index.md file was not being updated properly
Previously, the index.md file had previously existing fields left as is with an empty string, causing unintended behaviour where switching from dropdown to highlights would cause the dropdown to remain, due to it having higher priority.

The issue was fixed by removing all fields with empty strings when writing to the backend, such that the index.md file could be properly updated.