fix: title already in use bug
This commit fixes two things:
- stores the base api url as a state variable for reuse
- fixes the "title is already in use" bug which arises when trying
to rename a page to the same name.

For example, if I had a page called `new-page.md`, and I tried to
change the title to "New Page  " or if I added a few letters and then
removed them so that the title in the page settings modal reads
"New Page", I would get an error that says that the chosen title
is already in use. Indeed, it is already in use - but by the current
page which I am editing! This commit fixes that by excluding the current
page from the list of filenames to be compared against.