fix: redirect to page only if new page
This commit fixes the following bug: currently, even if just
editing the page settings of an existing page, the user is redirected
into the page to edit the content. However, this should only happen
if the page in question is a new page. This bug fixes that by
checking for `isNewPage` as part of the redirection conditions.