fix: clean-up after rebase
Previously, we had deleted the `EditResourcePage` component and subsumed
it under the `EditPage` component. However, upon rebase, the `EditResourcePage`
component was brought back. This repo re-deletes the `EditResourcePage` component
and also uses the updated `retrieveResourceFileMetadata` util function
and passes a `date` prop into the `SimplePage` component so that
the `EditPage` displays date for resource pages.