fix: update filename regex validation
Previously, the specified regex failed all cases since it did not
allow for characters before and after the `.` character respectively.

Since we only want a single `.` character in our filename, we need
to split on the `.` character and check that the resulting array length
is 2, to avoid cases where filenames like `a.jpg.jpg` pass the test.
This commit makes that check and adds a new test case for situations
where the filename is not specified but the extension is (for example,
`.jpg`).