Fix/more misc ui bugs (#168)
* chore: remove Menu option from sidebar

This commit removes the "Menu" tab from the CMS sidebar, since it
would take too much effort to reasonably implement the drag-and-drop
successfully for the beta release.

* fix: make the entire sidebar tab clickable

This commit resolves issue #158. Previously, the tabs in the sidebar
were only clickable on the tab, which is unintuitive for users. This
change modifies the code so that the entire tab is clickable.

* fix: remove edit button from page settings modal

Previously, one needed to click on the "Edit" button before editing
the permalink of a given page. This is inconsistent with the experience
of editing the title on the page settings modal.

This commit edits the page settings modal so that the "Edit" button
is removed, and the permalink is edited in the same way as the title.

This commit also ensures that the "Save" and "Delete" buttons are
disabled until the file SHA is received, so that we don't attempt to
make API calls which fail (it is impossible to update a file or delete
a file without its current SHA)

This commit fixes issue #166.

* feat: redirect to new page upon creation

This commit accomplishes a few things:
1. resolve issue #159 - upon creation of a new page, directly redirect
to edit that page instead of back to the Pages layout
2. remove the check for clashing permalinks

The decision to remove the check for clashing permalinks was made
because:
- Checking for clashing permalinks is inefficient - it requires
fetching every other file in the collection to check if there's a
permalink clash. This might be more effective when we switch to using
Github's GraphQL API.
- We don't currently perform any checks for clashing permalink in
Isomer, so we're still maintaining feature parity (not removing any
features that users currently have)

* fix: elaborate on filename validation

This commit breaks file name validation upon upload into two separate
steps:
1. a check for special characters
2. a check for proper file extensions

Previously, the two checks were made with the same regex expression,
which made it hard to provide a clear error message to the user. This
commit breaks them down so that users can have a clearer idea on how to
fix the error, with more elaborate error messages.

* fix: file name regex validation

This commit fixes the file name regex validation so that only
alphabets, numbers, dashes, and underscores are allowed in the
file name before the file extension.

* fix: title already in use bug

This commit fixes two things:
- stores the base api url as a state variable for reuse
- fixes the "title is already in use" bug which arises when trying
to rename a page to the same name.

For example, if I had a page called `new-page.md`, and I tried to
change the title to "New Page  " or if I added a few letters and then
removed them so that the title in the page settings modal reads
"New Page", I would get an error that says that the chosen title
is already in use. Indeed, it is already in use - but by the current
page which I am editing! This commit fixes that by excluding the current
page from the list of filenames to be compared against.

* fix: redirect to page only if new page

This commit fixes the following bug: currently, even if just
editing the page settings of an existing page, the user is redirected
into the page to edit the content. However, this should only happen
if the page in question is a new page. This bug fixes that by
checking for `isNewPage` as part of the redirection conditions.

* fix: filename validation regex

* fix: update filename regex validation

Previously, the specified regex failed all cases since it did not
allow for characters before and after the `.` character respectively.

Since we only want a single `.` character in our filename, we need
to split on the `.` character and check that the resulting array length
is 2, to avoid cases where filenames like `a.jpg.jpg` pass the test.
This commit makes that check and adds a new test case for situations
where the filename is not specified but the extension is (for example,
`.jpg`).

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>