Fix/unstack image upload modal (#169)
* fix: re-position "Select image" button

This commit accomplishes a few things:
- update media modal CSS to use rem instead of pixels
- reposition the "Select image" button in the media modal to the
bottom right corner using auto margins

* fix: media settings modal unable to close

This commit fixes the bug where we are unable to close the media
settings modal because the toggle state is not updated upon click
of the close button.

* feat: unstack media modals

Currently, in the EditPage layout, when you try to upload a photo,
a modal pops up within a modal - this is bad UI which might confuse
the users. The proposed solution is to close one modal when another one
is opened, that is, when someone clicks on "Upload photo", the original
media modal closes and the media setting modal opens. To do this, state
needs to be maintained on the parent EditPage layout instead of on the
media modal. This commit implements that fix.

Note: this change needs to be similarly applied to the EditResourcesPage
and EditCollectionPage as well.

* fix: shift "heading" button to start of markdown toolbar

* refactor: combine EditPage and EditResourcePage

Since most of the code between EditPage and EditResourcePage are duplicated,
we can add a few conditional statements to EditPage and delete EditResourcePage
altogether.

As part of this change, we need to pass an `isResourcePage` prop to EditPage so
that it knows whether to hit the pages endpoint, or the resources endpoint.

* fix: prevent image selection if no image selected

Currently, the MediaModal allows users to select an image using the `Select
${type}` button even if no image is selected. This results in an unhandled
rejection type error since we try to access a `path` attribute for an undefined
variable `selectedFile`. This commit fixes that by running the button callback
only if `selectedFile` exists.

* feat: clean up image/file modal design

This commit cleans up the image/file modal design in the following ways:
- uniform spacing between images across the modal
- padding the "Select File" header

* fix: position buttons on bottom right hand corner

This commit modifies the media settings modal and the resource settings
modal so that the save and delete buttons are located on the bottom right
hand corner, instead of the bottom left hand corner, for consistency.
This commit fixes issue #164.

* fix: clean-up after rebase

Previously, we had deleted the `EditResourcePage` component and subsumed
it under the `EditPage` component. However, upon rebase, the `EditResourcePage`
component was brought back. This repo re-deletes the `EditResourcePage` component
and also uses the updated `retrieveResourceFileMetadata` util function
and passes a `date` prop into the `SimplePage` component so that
the `EditPage` displays date for resource pages.

* fix: typo

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>