Feat/prevent navigation if unsaved changes (#170)
* refactor: Header component to use state change for back navigation

Currently, the Header component contains an <a> tag with a link to
the previous page. This commit refactors the header component so that
instead of using an <a> tag to redirect the user, we redirect using
React router based on a state change. Upon clicking the "Back to pages"
or "Back to site" button, a toggler toggles the shouldPerformBackNav
state and the component redirects backwards.

This change is made in preparation for an "Are you sure you want to
navigate away?" modal when users are in the `EditPage` component and
try to navigate away when they have unsaved changes.

* feat: add GenericWarningModal component

This commit adds the GenericWarningModal component, which performs in
a similar manner to the DeleteWarningModal, but allows the user to
specify the warning title and warning text.

This commit then modifies the Header component so that a user is
prompted to confirm that he/she wants to navigate away from the page
whenever they click the "Back to Pages" button, if the Header is part
of an EditPage.

* feat: trigger warning modal only if page content has changed

This commit introduces a shouldAllowEditPageBackNav prop to the Header component,
to be used in conjunction with the isEditPage prop. The
shouldAllowEditPageBackNav prop determines whether the warning modal
should be triggered for an EditPage layout.

In this case, the original markdown content is stored upon download
and used to compare with the updated markdown editor value. The value
of shouldAllowEditPageBackNav is thus true if the original content matches
the editor content, and false otherwise. This allows us to achieve the behavior
of preventing user from navigating away from the page if they have unsaved changes.
This works because everytime the user saves, the page is reloaded so the new content
is retrieved.

* fix: remove unused `type` prop for GenericWarningModal

This commit accomplishes a couple of things:
- it removes the unused `type` prop for the GenericWarningModal component
- it updates the proptypes for the GenericWarningModal to make `displayTitle` and
`displayText` required

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>